server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
mysql-connector-python